lstm_eigen_d logsumexp_grad code cleanup & lstm_eigen_types rename
