+++ src/cpp/modules/manualEigen/lstm_eigen_d_structures.h
@@ -2,7 +2,7 @@
 
 #include <vector>
 
-#include "lstm_eigen_helpers.h"
+#include "lstm_eigen_types.h"
 
 template<typename T>
 struct StateElementGradientModel
