+++ src/cpp/modules/manualEigen/lstm_eigen_types.h
